create local_players module
- rename ENGINE_MAX_PLAYERS to k_maximum_players
- rename ENGINE_MAX_LOCAL_PLAYERS to k_number_of_controllers
