+++ xlive/Blam/Engine/game/local_players.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "Blam/Cache/DataTypes/BlamPrimitiveType.h"
+
+#define k_number_of_controllers 4
+
+bool local_user_has_player(int user_index);
+datum local_user_get_player_idx(int user_index);
