+++ xlive/Blam/Engine/game/players.cpp
@@ -17,7 +17,7 @@ s_data_array* s_player::GetArray()
 
 bool s_player::IndexValid(int playerIndex)
 {
-	return playerIndex >= 0 && playerIndex < ENGINE_MAX_PLAYERS;
+	return playerIndex >= 0 && playerIndex < k_maximum_players;
 }
 
 s_player* s_player::GetPlayer(int playerIndex)
@@ -81,20 +81,6 @@ const wchar_t* s_player::GetName(int playerIndex)
 		return L"";
 	}
 	return GetPlayer(playerIndex)->properties[0].player_name;
-}
-
-bool local_user_has_player(int user_index)
-{
-	typedef bool(__cdecl* local_user_has_player_t)(int user_index);
-	auto p_local_user_has_player = Memory::GetAddress<local_user_has_player_t>(0x5139B);
-	return p_local_user_has_player(user_index);
-}
-
-datum local_user_get_player_idx(int user_index)
-{
-	typedef datum(__cdecl* local_user_get_player_idx_t)(int user_index);
-	auto p_local_user_has_player = Memory::GetAddress<local_user_get_player_idx_t>(0x5141D);
-	return p_local_user_has_player(user_index);
 }
 
 datum s_player::GetPlayerUnitDatumIndex(int playerIndex)
